% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_gipps.R
\name{simulate_gipps}
\alias{simulate_gipps}
\title{Simulate Gipps Model}
\usage{
simulate_gipps(
  resolution,
  N,
  dfn1,
  xn1,
  vn1,
  xn_first,
  vn_first,
  ln,
  an,
  Vn,
  tau,
  bn_const,
  bcap
)
}
\arguments{
\item{resolution}{Duration of a time frame. Typical values are 0.1, 0.5, 1.0 s. Double. Must match with the resolution of the observed lead vehicle data dfn1 defined below}

\item{N}{Number of Following Vehicles in the same lane (platoon). Integer.}

\item{dfn1}{Unquoted name of the dataframe that contains lead vehicle data.}

\item{xn1}{Name of the column in dfn1 that contains lead vehicle position. Character.}

\item{vn1}{Name of the column in dfn1 that contains lead vehicle speed. Character.}

\item{xn_first}{First value of vehicle position of each of the following vehicles. A list of doubles with size equal to N.}

\item{vn_first}{First value of vehicle speed of each of the following vehicles. A list of doubles with size equal to N.}

\item{ln}{Length of each of the lead vehicles. A list of doubles with size equal to N. Note that it is called as sn in Gipps Model}

\item{an}{Maximum acceleration which the driver wishes to undertake m/s2. Double.}

\item{Vn}{Desired speed/speed at which driver  wishes to travel m/s. Double.}

\item{tau}{Reaction Time s. Double.}

\item{bn_const}{Most severe braking that the driver wishes to undertake m/s2. Double and Negative.}

\item{bcap}{An estimate of lead vehicle deceleration m/s2. Double and Negative.}
}
\value{
A dataframe with lead and following vehicle(s) trajectories
}
\description{
This function takes in the lead vehicle trajectory and calculates speed, spacing and acceleration of the following vehicle using the Gipps Model.
}
\examples{
# Time
last_time <- 3000 ## s
time_frame <- 0.1 ## s
Time <- seq(from = 0, to = last_time, by = time_frame)
time_length <- length(Time)



## Lead vehicle
vn1_first <- 13.9 ## first speed m/s
xn1_first <- 100 ## position of lead vehicle front center m
bn1_complete <- c(rep(0, 29500),
                 rep(-5, time_length - 29500))



#############################################
### Complete speed trajectory of Lead vehicle
#############################################

vn1_complete <- rep(NA_real_, time_length) ### an empty vector
xn1_complete <- rep(NA_real_, time_length) ### an empty vector

vn1_complete[1] <- vn1_first
xn1_complete[1] <- xn1_first

for (t in 2:time_length) {

 ### Lead vehicle calculations
 vn1_complete[t] <- vn1_complete[t-1] + (bn1_complete[t-1] * time_frame)

 vn1_complete[t] <- ifelse(vn1_complete[t] < 0, 0, vn1_complete[t])


xn1_complete[t] <- ifelse(
 vn1_complete[t] > 0,
 xn1_complete[t-1] + (vn1_complete[t-1] * time_frame) +
   (0.5 * bn1_complete[t-1] * (time_frame)^2),
 xn1_complete[t-1]
)

}



ldf <- data.frame(Time, bn1_complete, xn1_complete, vn1_complete)

## Run the Gipps function:
simulate_gipps(

############## Simulation Parameters #######################
resolution=0.1,
N=5,


############### Lead Vehicle Data #########################
dfn1=ldf,
xn1="xn1_complete",
vn1="vn1_complete",


############### Following Vehicle Data ####################
xn_first=list(85, 70, 55, 40, 25),
vn_first=list(12, 12, 12, 12, 12),
ln=list(6.5, 6.5, 6.5, 6.5, 6.5),


############### Model Parameters ##########################
an=2,
Vn=14.4,
tau=0.1,
bn_const=-1.5,
bcap=-2
)
}
